[jay.den] step3 PR 요청합니다. (#65)
* refactor : 예외관련 수정

- 예외명, DTO 클래스명 등 일부 네이밍 수정

- ReservationResponse 생성자 추가
before : 모든 필드를 입력하는 생성자
after : Reservation 을 입력으로 받아 생성하는 생성자 추가

- ReservationController 조회 response 수정
before : 다른 api 와 달리 Response 객체를 return
after : 다른 api 와의 일관성을 위해 ResponseEntity 를 return 하도록 수정

- DAO 의 findById 의 return 값을 Optional 로 수정
before : Reservation 을 return 하고 null 인지 확인하는 방식으로 예외처리
after : 두 DAO 구현체에서 조회 결과로 List 를 생성하고, List 의 stream.findAny 로 결과를 찾아 Optional 로 return 하도록 하여 NullPointerException 방지. Optional 의 isPresent 로 결과 확인하여 예외처리

- Enum ErrorCode 추가 및 Exception, Handler 수정
before : 중복 예약에 대해서만 Exception 을 만들고 해당 예외 발생시 메시지 보냄
after : 여러개의 예외가 추가될 가능성을 고려하여 Enum ErrorCode 를 정의하고, 해당 ErrorCode 를 담는 Exception 으로 변경. Handler 에서는 Exception 에서 HttpStatus 와 Message 를 받아 return 하도록 수정.

* refactor : 중복코드 제거

- mapper 사용
before : DAO interface 구현한 두 클래스에서 mapping 시 중복코드 다수 존재
after : 중복되는 코드부분을 RowMapper 와 PreparedStatementCreator 로 interface 에 default 로 구현하여 중복코드 제거

- ReservationJdbcDao 메소드 분리
before : 각 메소드에서 db 와 connection 을 만들고 닫는 부분의 코드가 중복됨
after : 메소드 분리하여 중복 제거함

- ReservationService.delete 예외처리
before : 삭제하려는 id 가 존재하지 않아도 별도 코멘트 하지 않음
after : 삭제하려는 id 가 존재하지 않으면 예외발생시켜 처리되지 않음을 응답할 수 있도록 함

* refactor : 콘솔 애플리케이션 예외처리

발생되는 예외 catch 하여 출력하도록 예외처리 수정함

* refactor : TimeTable 추가 및 예외처리 추가

- TimeTable
before : 시간표 없이 날짜/시간 중복만 확인
after : 시간표 적용하여 없는 시간에 대한 예외처리

- 예외처리
before : 요청 파라미터의 유효성에 대한 검증 없음
after : 모델에 검증로직 구현하여 요청 파라미터의 유효성 검증 추가

- Test
before : 입력 파라미터, 등록되지 않은 id 등에 대한 테스트 없음
after : 테스트 추가

* refactor : 테스트 메소드명 수정 및 Reservation 단위테스트 추가

* refactor : theme 관리 기능 관련 DDL 쿼리 변경 및 도메인에 id 추가

* feat : Theme 단위테스트 추가

- Theme 의 name 과 desc 와 price 는 입력이 필수적이므로 잘못된 입력의 예외처리를 위해 단위테스트 및 유효성 검사 로직 추가

* feat : 테마 생성 api 구현

- ThemeController, ThemeService, ThemeDao, ThemeJdbcTemplateDao 생성
- 테마 생성 api 구현
- 이미 등록된 테마 이름이 있는 경우 예외발생하도록 처리

* feat : 테마 조회 api 구현

- 테마 단일 조회 구현
유효하지 않은 id 로 조회시 예외처리

- 테마 목록 조회 구현

* feat : 테마 수정 api 구현

- uri path variable 로 id, body 로 변경사항 담아 put 요청시 수정되도록 구현
- 유효하지 않은 id (0 이하) 에외처리
- 유효하지만 db 에 존재하지 않는 id 로 요청시 create 처리
- 유효하고 db 에 존재하지만 다른 레코드에 동일한 테마 이름이 있으면 예외처리

* feat : 테마 삭제 api 구현

- uri path variable 로 id 받아 삭제하도록 구현
- 유효하지 않은 id (0 이하) 에외처리 : 메서드 분리
- 유효하지만 db 에 존재하지 않는 id 로 요청시 예외처리

* fix : Theme 추가에 따른 Reservation(web) 로직 수정

- 예약
* 입력값 수정
날짜, 시간, 이름 -> before + 테마 id
* 테마 id 유효성 검사
* 테마 id 에 해당하는 테마가 등록되어있지 않으면 예외처리
* 예약 중복 검사시 날짜, 시간 + 테마 까지 확인하여 테마별로 분기

- 조회
* 조회시 테마 id 로 테마 테이블을 join 하여 예약 내역에 테마 내용이 포함되도록 처리

* feat : 도메인 각 JdbcDao 로직 수정 및 작성

- 예약 관련 로직 수정에 따라 콘솔에서도 작동하도록 하기 위해 ReservationJdbcDao 로직 수정
- 테마 관리를 위한 콘솔 애플리케이션 로직 수정은 계획에 없으나, 애플리케이션의 작동을 위해 ThemeJdbcDao 도 작성

* refactor : 예외처리 추가

- 특정 테마에 대해 예약된 건이 존재할 경우 테마 수정 또는 삭제가 불가능하도록 예외처리 추가

* refactor : 통합테스트시 테스트 격리부분 분리

before : 통합테스트마다 테스트 격리를 위해 사용된 db 를 초기화하기 위한 코드를 따로 작성해주어야 했음. 이를 위해 추가적으로 jdbcDao 등을 세팅해주어야 했으며 테스트를 하기 위해 db 를 초기화하는 코드를 각 dao 에 작성해야 했음.
after : 통합테스트시 각 테스트마다 테스트가 완료되면 db 를 초기화할 수 있도록 TestExecutionListener 를 구현 및 등록함. 또한 커스텀 어노테이션을 통해 SpringBootTest 와 TestExecutionListener 등록을 묶어 결과적으로 IntegrationTest 어노테이션만 사용하면 적용되도록 함.

+++ src/main/java/nextstep/service/ReservationService.java
@@ -2,55 +2,71 @@ package nextstep.service;
 
 import nextstep.domain.Reservation;
 import nextstep.domain.Theme;
-import nextstep.dto.ReservationRequestDto;
-import nextstep.dto.ReservationResponseDto;
-import nextstep.exceptions.exception.DuplicatedDateAndTimeException;
+import nextstep.dto.ReservationRequest;
+import nextstep.dto.ReservationResponse;
+import nextstep.exceptions.ErrorCode;
+import nextstep.exceptions.exception.InvalidRequestException;
 import nextstep.repository.ReservationDao;
+import nextstep.repository.ThemeDao;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.Optional;
 
 @Service
 public class ReservationService {
     private final ReservationDao reservationDao;
+    private final ThemeDao themeDao;
 
-    public ReservationService(@Qualifier("reservationJdbcTemplateDao") ReservationDao reservationDao) {
+    public ReservationService(
+            @Qualifier("reservationJdbcTemplateDao") ReservationDao reservationDao,
+            @Qualifier("themeJdbcTemplateDao") ThemeDao themeDao
+    ) {
         this.reservationDao = reservationDao;
+        this.themeDao = themeDao;
     }
 
-    public Long reserve(ReservationRequestDto reservationRequestDto) {
-        LocalDate date = reservationRequestDto.getDate();
-        LocalTime time = reservationRequestDto.getTime();
+    public Long reserve(ReservationRequest reservationRequest) {
+        LocalDate date = reservationRequest.getDate();
+        LocalTime time = reservationRequest.getTime();
+        Long themeId = reservationRequest.getThemeId();
 
-        if (reservationDao.countByDateAndTime(date, time) > 0) {
-            throw new DuplicatedDateAndTimeException();
+        if (reservationDao.countByDateAndTimeAndThemeId(date, time, themeId) > 0) {
+            throw new InvalidRequestException(ErrorCode.RESERVATION_DUPLICATED);
         }
 
-        Theme theme = new Theme("워너고홈", "병맛 어드벤처 회사 코믹물", 29_000);
-        Reservation newReservation = new Reservation(date, time, reservationRequestDto.getName(), theme);
+        validateId(themeId);
+        Optional<Theme> themeFound = themeDao.findById(themeId);
+        if (themeFound.isEmpty()) {
+            throw new InvalidRequestException(ErrorCode.THEME_NOT_FOUND);
+        }
 
+        Reservation newReservation = new Reservation(date, time, reservationRequest.getName(), themeFound.get());
         return reservationDao.save(newReservation);
     }
 
-    public ReservationResponseDto retrieve(Long id) {
-        Reservation reservation = reservationDao.findById(id);
-        if (reservation != null) {
-            return new ReservationResponseDto(
-                    reservation.getId(),
-                    reservation.getDate(),
-                    reservation.getTime(),
-                    reservation.getName(),
-                    reservation.getTheme().getName(),
-                    reservation.getTheme().getDesc(),
-                    reservation.getTheme().getPrice()
-            );
+    public ReservationResponse retrieve(Long id) {
+        validateId(id);
+        Optional<Reservation> reservationFound = reservationDao.findById(id);
+        if (reservationFound.isEmpty()) {
+            throw new InvalidRequestException(ErrorCode.RESERVATION_NOT_FOUND);
         }
-        return null;
+        return new ReservationResponse(reservationFound.get());
+    }
+
+    private void validateId(Long id) {
+        if (id == null || id <= 0) {
+            throw new InvalidRequestException(ErrorCode.INPUT_PARAMETER_INVALID);
+        }
     }
 
     public void delete(Long id) {
+        validateId(id);
+        if (reservationDao.findById(id).isEmpty()) {
+            throw new InvalidRequestException(ErrorCode.RESERVATION_NOT_FOUND);
+        }
         reservationDao.delete(id);
     }
 }
